% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBioYear.R
\name{addBioYear}
\alias{addBioYear}
\title{Add Biological Year to data}
\usage{
addBioYear(df, bioStart)
}
\arguments{
\item{bioStart}{biological year start date, in day of year format \code{\link{yday}}}
}
\value{
a dataframe with columns:
\describe{
  \item{bioYear}{Biological year corresponding to user-defined start day}
}
}
\description{
This function adds a column designating biological year starting on a user-designated date 
e.g. 1 June (day of year 152) for caribou
}
\examples{
data("barrenground")

head(barrenground)

# bioStart is yday 152 (01 June)

barrenground <- addBioYear(barrenground, bioStart = 152)
}
