% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{barrenground}
\alias{barrenground}
\title{Example of movement data for 4 Barren-ground caribou individuals}
\format{
data frame containing the ID, the time of the location, the Lon and Lat coordinates (WGS84), 
the x and y coordinates (in Lambert Conformal Conic as: "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs")
}
\source{
GNWT
}
\usage{
data(barrenground)
}
\description{
Example of movement data for 4 Barren-ground caribou individuals
}
\examples{
data(barrenground)


}
\keyword{data}
