% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{bathurst_lines}
\alias{bathurst_lines}
\title{Some spring migrations of Bathurst caribou}
\format{
Simple feature lines, with ID and Year (all 2010)
}
\source{
GNWT
}
\usage{
data(bathurst_lines)
}
\description{
Some spring migrations of Bathurst caribou
}
\examples{
data(bathurst_lines)
mapview(bathurst_lines[,"ID"])

}
\keyword{data}
