% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{caribou}
\alias{caribou}
\title{Movement data of barren-ground caribou}
\format{
data frame with the following columns: 
#' \describe{
  \item{ID}{ID of animal}
  \item{sex}{Sex of the animal}
  \item{Time}{Date and time of each GPS location}
  \item{Year}{Year of the GPS location}
  \item{Lon,Lat}{Coordinates of the GOS locations in WGS84}
  \item{x,y}{Coordinates of the GOS locations in Canada Lambert Conformal Conic (crs = "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs")}
}

These are real data courtesy of NWT Department of Environment and Natural Resources
but anonymized and shifted in time, longitude and latitude.
}
\source{
GNWT Department of Environment and Natural Resources
}
\usage{
data(caribou)
}
\description{
Movement data of barren-ground caribou
}
\examples{
data(caribou)
head(caribou)

# plotting locations on a base graphics map
require(mapdata)
maps::map("worldHires", xlim = c(-135,-120), ylim=c(62,67), col = "grey")
ids <- levels(caribou$ID)
for(i in 1:length(ids))
  with(caribou \%>\% subset(ID == ids[i]), lines(Lon, Lat, col = i))

# plotting points with mapview
caribou.sf <- st_as_sf(caribou \%>\% mutate(Lat = Lat, Lon = Lon), 
                       coords = c("Lon","Lat")) \%>\% st_set_crs(4326)
mapview::mapview(caribou.sf[,"ID"])

# converting to lines
lines <- caribou.sf \%>\% group_by(ID) \%>\% 
  summarize(do_union=FALSE) \%>\% 
  st_cast("LINESTRING")
mapview::mapview(lines[,"ID"])
}
\keyword{data}
