% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPtsInPolygons.R
\name{countPtsInPolygons}
\alias{countPtsInPolygons}
\title{Count Points in polygons}
\usage{
countPtsInPolygons(sf, bounds, id_col, assigned_col = "study_site")
}
\arguments{
\item{sf}{simple feature points with columns:
ID: designating individuals
Time: Date and time in POSIXct}

\item{bounds}{simple features multipolygon  with column:}

\item{assigned_col:}{study area an animal was assigned to (likely from movebank)}

\item{bounds_col:}{identifying unique polygon boundaries}
}
\value{
a data frame with columns:
ID: animal ID
start: first date of monitoring
end: last date of monitoring
n_pts: total number of locations
assigned: study area animal was originally assigned
columns for each polygon boundary with number of locations
}
\description{
This function takes each animal ID and counts how many of its locations are 
in polygon boundaries. This is helpful to help decide whether the study area 
an animal was assigned to is accurate or to identify which region an animal 
spends most of its time.
}
\examples{
require(TuktuTools)
data(caribou)

# Winter (December 1 to March 15)
winterpoints <- subset(caribou, (month(Time) == 12 & Year == 2003) |
                         (month(Time) \%in\% c(1,2) & Year == 2004) |
                         (mday(Time) < 16 & month(Time) == 2 & Year == 2004)) \%>\% droplevels

# aggregate to keep one point per week to avoid autocorrelation
winterpoints <- winterpoints \%>\% mutate(week = week(Time)) \%>\% group_by(week) \%>\%
  summarize(x = mean(x), y = mean(y), ID = unique(ID)) \%>\% ungroup \%>\% as.data.frame

# convert to simple features in metric system
CRS = "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
winter.sf <- st_as_sf(winterpoints, coords = c("x","y"), crs=CRS)

# estimate MCP 95\% and 50\%
mcp95 <- getMCP(winter.sf)
mcp50 <- getMCP(winter.sf, percent = 50)

# Visualize ranging areas
require(mapview)

# Minimum convex polygon
mapview(winter.sf) +
  mapview(mcp95, col.regions = "red") +
  mapview(mcp50, col.regions = "darkred")
}
