% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutperiod.R
\name{cutperiod}
\alias{cutperiod}
\title{Cut a timeseries for the period of interest}
\usage{
cutperiod(
  df,
  start,
  end,
  nfixes,
  dayloss,
  CRS =
    "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
)
}
\arguments{
\item{df}{a data frame containing columns: ID as individual identifiant,
x and y: relocations of individuals (in N Canada Lambert Conformal Conic)
Time: date and time vector of the relocation (of class POSIXct)}

\item{start}{starting month and day of the period of interest as a vector c(mm,dd)
(example for the 19th of May: c(05,19))}

\item{end}{end month and day of the period of interest as a vector c(mm,dd)
(example for the 7th of July: c(07,07))}

\item{nfixes}{minimun number of fixes per day
(taking the average number of fixes per day for each individual across the period of interest)}

\item{dayloss}{maximum number of days with missing locations
(for example, if an individual has loss signal for more than 3 consecutive days, it will be excluded from the dataset)}

\item{CRS}{the coordinates projection (default is Canada Lambert Conformal Conic centered on the area from the Western Arctic to the Hudson Bay:
"+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs")}
}
\value{
a dataframe containing only concerned period and individuals that match the defined rules
}
\description{
Function to prepare and clean the data by keeping the period of interest.
We first have to keep GPS locations for the period of interest:
for example, for the calving period:
Following Cameron et al. 2018, we defined the calving period from
May 19 to July 7
In addition, the parturition model seems quite accurate, but if some individuals
have lost signal for several days it could lead to a undetermination
of the calving event.
}
\details{
Before running the parturition model, we have to establish rules
to determine the individuals that will be included in the analysis
and the ones that will be excluded
For example:
-exclude individuals with less than 1 fix per day
-exclude individuals that lost signal for more than 3 days
-exclude individuals for which monitoring stopped before July 15 or began after May 19
}
