% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadWeather.R
\name{downloadWeather}
\alias{downloadWeather}
\title{Download Dailytemp}
\usage{
downloadWeather(station, dir = tempdir(check = TRUE))
}
\arguments{
\item{{station}}{station ID}

\item{{dir}}{folder to save the file in (default: temporary directory)}
}
\description{
Given a known station ID, this function automatically downloads the corresponding daily temperature file 
from "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/gsn/" and converts the relevant average, maximum, 
minimum daily temperature into a dataframe.
}
\examples{
mydata <- downloadWeather("AO000066160")
}
\references{
The journal article describing GHCN-Daily is: 
Menne, M.J., I. Durre, R.S. Vose, B.E. Gleason, and T.G. Houston, 2012:  An overview 
of the Global Historical Climatology Network-Daily Database.  Journal of Atmospheric 
and Oceanic Technology, 29, 897-910, doi:10.1175/JTECH-D-11-00103.1.

To acknowledge the specific version of the dataset used, please cite:
Menne, M.J., I. Durre, B. Korzeniewski, S. McNeal, K. Thomas, X. Yin, S. Anthony, R. Ray, 
R.S. Vose, B.E.Gleason, and T.G. Houston, 2012: Global Historical Climatology Network - 
  Daily (GHCN-Daily), Version 3. [indicate subset used following decimal, 
                                    e.g. Version 3.12]. 
NOAA National Climatic Data Center. http://doi.org/10.7289/V5D21VHZ
}
