% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCalving.R
\name{estimateCalving}
\alias{estimateCalving}
\title{Parturition timing function}
\usage{
estimateCalving(
  df,
  int,
  kcons,
  models = c("full", "calfonly"),
  drawplot = FALSE,
  saveplot = FALSE
)
}
\arguments{
\item{df}{a dataframe containing the movement rate between subsequent 
relocation of individuals, as obtained using `getSpeed`.
See ?get.speed for more information on the Data requirements}

\item{int}{integer value indicating the minimum number of days between
the beginning of the time series and the first BP (calf birth),
the last BP (calf death) and the end of the time series,
and between the two BPs.
The main reason for this constraint is that a minimum number of data points are required in
each section of the time-series to be able to estimate model parameters.
We recommended 9 relocations, thereby 3 days.}

\item{kcons}{vector of the minimum and maximum time it takes the female to recover normal speed (in days)}

\item{models}{either "full" to fit all three models (i.e., no calf, calf and calf death models), or 
"calfonly" to fit only the no calf and calf models.}

\item{drawplot}{if TRUE, the function will draw plots of the movement rate in function to the date
with the prediction line of the best model selected by AIC, as well as the path of the individual 
with the calving location and a plot of the net square displacement 
in function of time with  a vertical line corresponding to the estimated calving date}
}
\value{
returns list of 3 dataframes:
`statistics` containing the best model for each female, selected based on AIC, the AIC of the 3 models 
(i.e. no calf model, calf model and calf death model) and the negative likelihood 
of each model.
`Par` containing the estimated parameters of all the models (0: no calf, 1: calf, 2: calf death)
`results` containing the best model, for each female, based on AIC,
 the calving timing and location (if the calf model is the best model),
and the mortality timing and location of the calf (if the calf death is the best model).
}
\description{
Function that determine calving status of a female (i.e. no calf, with a calf, calf lost)
the calving date and the calf death date (if any).
This is an adaptation of the individual based method developed by Demars et al. (2013),
which has proven good reliability to estimate calving status and calving date for females
from the Western Arctic Herd in a previous study (Cameron et al. 2018).
However, we adapted this method to be able to infer parturition based on the 
female movement rate through time.
}
\details{
All models assume speed follows a Gamma distribution and differ in two parameters: shape
and scale, which correspond to \eqn{mean(speed)^2 / var(speed)}
and \eqn{var(speed) / mean(speed)}, respectively.
The mean speed is thus equal to \eqn{shape * scale}.
\itemize{
\item For the model representing females that do not calve: the mean speed remains constant
through the entire calving period.
\item For the model representing females that had a calf who survived 4 weeks after birth:
the mean speed is constant before calving, then abruptly drops for calving, creating a break point.
After calving, the mean speed increases progressively following:
\eqn{ (shape.calving * (mean(shape) - shape.calving) / k * time) * (scale.calving * (mean(scale) - scale.calving) / k) * time)}
where k is the time, defined in days, required for the calf to achieve adult movement rates.
\item For the model representing females losing calves: there is an abrupt change in the slope of
the post-calving increase, creating a second break point after which the mean speed immediately
recovers its pre-calving value.}
The models therefore differ in their number of parameters to estimate:
the no calf model has two – shape and scale;
the calf model has four – shape, scale, k and calving date;
and the calf death model has five – shape, scale, k, calving and calf death.
We discriminated among models using Akaike’s Information Criterion (AIC)
with the best model being the one with the lowest AIC value.

The visualization portion of this code is adapted directly from Matt Cameron and supplementary materials in Cameron et al. 2018.
}
\examples{
require(TuktuTools)

data(caribou)

# remove potential outliers and keep only females
caribou.cleaned <- removeOutliers(caribou) \%>\% subset(outlier == "FALSE" & sex == "f")

# keep data during calving period (i.e., May 19 to July 7), no more than 3 consecutive days of missing data
# and at least 1 fix per day
caribou.prepped <- prepData(caribou.cleaned, start = "05-19", end = "07-07", 
                            nfixes = 1, dayloss = 3)

# get movement rate for the females
caribou.mr <- getSpeed(caribou.prepped)
head(caribou.mr)

# estimate calving with the no calf and calf model
parturitions <- estimateCalving(caribou.mr, int=3, kcons =c(5,21), models = "full", drawplot = TRUE)
head(parturitions)

# get plot for one individual
table(caribou.mr$ID, caribou.mr$Year) # take Vixen in 2008

# what is the identified model for Vixen in 2008?
parturitions$results[parturitions$results$ID == "Vixen" & 
                       parturitions$results$Year == 2008,] # "calfdeath"

plotCalvingRange(df = caribou.mr[caribou.mr$ID == "Vixen" & 
                                   caribou.mr$Year == 2008,], int = 3,
               kcons = c(5,21), bestmodel = "calfdeath")
}
\references{
DeMars, C., M. Auger-Méthé, U. Schlägel, S. Boutin, (Published online) Inferring Parturition and Neonate Survival from Movement Patterns of Female Ungulates. Ecology and Evolution. DOI: 10.1002/ece3.785
Cameron, M.D., Joly, K., Breed, G.A., et al. (2018). Movement-based methods to infer parturition events in migratory ungulates. Canadian Journal of Zoology. 96:1187–1195. https://doi.org/10.1139/cjz-2017-0314
}
