% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspeed.R
\name{get.speed}
\alias{get.speed}
\title{get speed}
\usage{
get.speed(
  df,
  CRS =
    "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
)
}
\arguments{
\item{df}{a data frame containing columns: ID_Year as individual identifiant per year,
x and y: relocations of individuals (in N Canada Lambert Conformal Conic)
DateTime: date and time vector (of class POSIXct)}

\item{CRS}{the coordinates projection (default is Canada Lambert Conformal Conic:
"+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs")}
}
\value{
The function returns the same dataframe than 'df' with columns speed, which is the speed between subsequent relocations (in m.h-1),
time.mid corresponding to the average time of subsequent relocations,
dhours which is an index of the cumulative hours of each speed values from the first speed value,
dt the time lags between subsequent relocations (in hour)
}
\description{
This function calculate speed from raw movement data containing x and y coordinates (in meters) with corresponding date and time
in POSIXct format. This will calculate the speed between subsequent relocations (i.e. of each step), as well as the middle time
of the step.
}
