% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBasemap.R
\name{getBasemap}
\alias{getBasemap}
\title{get base map raster}
\usage{
getBasemap(
  gg.ext,
  map_service = "osm",
  map_type = "terrain_bg",
  map_res,
  m.crs = NULL,
  map_dir = "basemap",
  delete_tiles = TRUE
)
}
\arguments{
\item{map_service}{workd best with the default "osm" (Open street map)}

\item{map_type}{see options in \code{\link{moveVis::get_maptypes()}}}

\item{m.crs}{Coordinate reference system}

\item{map_dir}{a location to (temporarily) store basemap tiles}

\item{delete_tiles}{whether or not to clear the basemap directory.}
}
\description{
Based on the (hidden) getMap function in moveVis, with some slight modifications for 
steamlining.
}
\examples{
require(moveVis)
require(TuktuTools)
require(slippymath)
require(pbapply)
require(curl)
require(magick)

data("caribou")
b <- caribou \%>\% group_by(ID) \%>\% st_as_sf(coords=c("Lon", "Lat"), crs=4326) \%>\%
  summarize(do_union=FALSE) \%>\% st_cast("LINESTRING")

gg.ext <- st_bbox(b)
dir.create("basemap")

a <- getBasemap(gg.ext = st_bbox(b), map_service = "osm", map_type = "terrain_bg", 
                map_dir ="basemap", map_res = .5, m.crs = st_crs(b))

plotRGB(a)
plot(st_geometry(b \%>\% st_transform(crs(a[[1]]))), add = TRUE, col = 1:nrow(b))
}
