% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCalvingPeriod.R
\name{getCalvingPeriod}
\alias{getCalvingPeriod}
\title{Bivariate segmentation of the daily movement rate and daily ranging area}
\usage{
getCalvingPeriod(df, drawplot = TRUE, min.duration = 5, ...)
}
\arguments{
\item{df}{a data frame containing columns: the daily ranging area (area), movement rate (speed), day of year (yday) and date (Date)}

\item{drawplot}{whether to visualize the segmentation}

\item{min.duration}{minimum duration of segment}

\item{...}{additional parameters to pass to plot function (e.g. xlim and ylim)}
}
\value{
a data frame with the dates of the calving period for the year
}
\description{
Function to identify the annual calving period at the population level.
The function segmentates a joint time series of daily herd movement rate and ranging area.
The function uses a two-dimensional time-space clustering method on the bivariate time series 
implemented in the segclust2d R package (Patin et al., 2020). 
The segmentation is performed on the square root of the area, which has a more normal distribution 
compared to raw area. The segmentation considers 5 days as the minimum number of days per period, 
on 3 different periods: pre-calving, calving, post-recovery.
}
\references{
Patin, R., Etienne, M., Lebarbier, E., Chamaillé‐Jammes, S., & Benhamou, S. (2020). 
Identifying stationary phases in multivariate time series for highlighting behavioural modes and 
home range settlements. Journal of Animal Ecology, 89(1), 44–56. https://doi.org/10.1111/1365-2656.13105
}
