% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDailyRange.R
\name{getDailyRange}
\alias{getDailyRange}
\title{get Daily Range}
\usage{
getDailyRange(sf, crs, method = c("LoCoH", "KernelUD"), ...)
}
\arguments{
\item{sf}{a simple feature containing (daily) GPS coordinates used to estimate 
the (daily) ranging area}

\item{crs}{the crs of the coordinates in sf}

\item{method}{either "LocoH" or "KernelUD", the method used to estimate the daily ranging area (see getLoCoH and getKernelUD for more details)}

\item{...}{additional parameters to pass to \link{kernelUD} for the "KernelUD" method or \link{lhs.iso.add} for the "LoCoH" method}

\item{nn}{(if method is "LoCoH") the number of nearest neighbors used to construct the local convex polygons}

\item{level}{(if method is "LoCoH") A numeric value of the level of the isopleths (value comprised between .05 and 1, default is .95)}

\item{percent}{(if method is "KernelUD") percent kernel (value comprised between 5 and 100, default is 95)}

\item{grid}{(if method is "KernelUD") grid points (default 200, higher than adehabitat's)}
}
\value{
The function returns a simple feature of the isopleth(s), with the area
}
\description{
This function estimate (daily) Local Convex Hulls (LoCoH) by converting a simple feature into a LoCoH-xy
object, identifying the nearest neighbors, creating a LoCoH-hullset object by constructing the 
local convex polygon using the nearest neighbors (nn) of each data point
}
\examples{
require(TuktuTools)

data(simulated_migrations)

head(simulated_migrations)

# for this example, we will take all the locations of the animals at a given day, 
# whatever the year, to have enough locations for this day.
# However, this method has been developed to estimate daily ranges for animals 
# monitored the same year.

# add the day of year
migrants <- simulated_migrations.sf

# get average daily location for each individual and year then assign the same year for computing the daily range
migrants.dailymean <- simulated_migrations.sf \%>\% getDailyMean

# daily KernelUD
require(adehabitatHR)
dailyranges.ud <- getDailyRange(migrants.dailymean, method = "KernelUD", 
                                percent = 95) \%>\%
  mutate(Date = as.Date(yday, origin = "2014-12-31"))

mapview::mapview(st_as_sf(dailyranges.ud), zcol = "yday")
with(dailyranges.ud, plot(Date, area, type = "l"))

}
