% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHerdAssignment.R
\name{getHerdAssignment}
\alias{getHerdAssignment}
\title{Get herd assignment using spatial clustering}
\usage{
getHerdAssignment(df, K, plotIt = TRUE)
}
\arguments{
\item{df}{a dataframe containing x and y coordinates and the ID_year
(Identifiant of individuals for each year of survey) of all datasets combined in one}

\item{K}{number of herds (can be determined using getKnumber function in this package)}

\item{plotIt}{if TRUE, plot the GPS location with herd assignment}
}
\value{
the dataframe with GPS location and a cluster column with
a number comprised between 1 and K, corresponding to the herd assignment.
}
\description{
Function to determine herd assignment using K-means analysis on GPS locations of individuals,
using the average location for each individual and year.
}
\references{

}
