% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKnumber.R
\name{getKnumber}
\alias{getKnumber}
\title{Get the number of herd (i.e. populations) from GPS data}
\usage{
getKnumber(df, kmin, kmax, method = "kmeans", index = "all")
}
\arguments{
\item{df}{a dataframe containing x and y coordinates and the ID_year
(Identifiant of individuals for each year of survey)}

\item{kmin}{minimum number of herds}

\item{kmax}{maximum number of herds}

\item{method}{the cluster analysis method to be used among: "ward.D", "ward.D2",
"single", "complete", "average", "mcquitty", "median", "centroid", "kmeans".
Default is "kmeans".}

\item{index}{the index to be calculated. This should be one of : "kl", "ch", "hartigan",
"ccc", "scott", "marriot", "trcovw", "tracew", "friedman", "rubin", "cindex", "db",
"silhouette", "duda", "pseudot2", "beale", "ratkowsky", "ball", "ptbiserial", "gap",
"frey", "mcclain", "gamma", "gplus", "tau", "dunn", "hubert", "sdindex", "dindex", "sdbw",
"all" (all indices except GAP, Gamma, Gplus and Tau),
"alllong" (all indices with Gap, Gamma, Gplus and Tau included)}
}
\value{
The function returns a plot of the Number of indices proposing the given optimal number
of herds, the optimal number of clusters proposed by each index as well as the value of each
index and the Optimal Cluster assignment based on the consensus on the optimal number of cluster.
}
\description{
Function to choose the optimal number of herds (i.e. populations),
based on GPS data of individuals using NbClust function from the NbClust package.
This function provides 30 indices to determine the optimal number of clusters to be used.
See ?NbClust for more information.
}
\references{
Charrad M., Ghazzali N., Boiteau V., Niknafs A. (2014).
NbClust: An R Package for Determining the Relevant Number of Clusters in a Data Set.
Journal of Statistical Software, 61(6), 1-36.URL http://www.jstatsoft.org/v61/i06/.
}
