% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLoCoH.R
\name{getLoCoH}
\alias{getLoCoH}
\title{get LoCoH and get Daily Range}
\usage{
getLoCoH(
  sf,
  nn = 10,
  level = 0.9,
 
    CRS = "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs",
  ...
)
}
\arguments{
\item{sf}{a simple feature containing (daily) GPS coordinates used to estimate 
the (daily) ranging area}

\item{nn}{the number of nearest neighbors used to construct the local convex polygons}

\item{level}{A numeric value of the level of the isopleths}

\item{CRS}{the CRS of the simple feature (default is "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs")}
}
\value{
The function returns a simple feature of the isopleth(s), with the area
}
\description{
This function estimate (daily) Local Convex Hulls (LoCoH) by converting a simple feature into a LoCoH-xy
object, identifying the nearest neighbors, creating a LoCoH-hullset object by constructing the 
local convex polygon using the nearest neighbors (nn) of each data point
}
