% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPairwiseOverlaps.R
\name{getPairwiseOverlaps}
\alias{getPairwiseOverlaps}
\title{Compute pairwise overlaps}
\usage{
getPairwiseOverlaps(sf.list, column = "ID", method = "VI", ...)
}
\arguments{
\item{sf.list}{list of simple feature tracks}

\item{column}{column by which to split the data}

\item{method}{method of kernel overlap}

\item{...}{additional parameters for \code{\link{kerneloverlap}} function - notably "grid"}
}
\value{
data frame with ID1, ID2 and overlap index.
}
\description{
This function (rather quickly) computes pairwise overlaps between daily mean locations
of animals.  By default it uses the "volume intersection" index (see \code{\link{adehabitatHR::kerneloverlap}}).
}
