% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeatherAtLocation.R
\name{getWeatherAtLocation}
\alias{getWeatherAtLocation}
\title{Get Weather at Location}
\usage{
getWeatherAtLocation(
  longitude,
  latitude,
  start,
  end,
  threshold = 200,
  dir = tempdir(check = TRUE),
  keep.raw = FALSE
)
}
\arguments{
\item{start}{first year of observation.  The function will keep widening its 
search radius until it finds a station with data reaching back to the startyear.  
Also, the returned data will only reach back to the start year.}

\item{end}{last year of observations.}

\item{threshold}{distance within which to limit search for weather stations (default 200 km)}

\item{dir}{directory to store the raw observations}

\item{keep.raw}{whether or not to keep the raw file}

\item{{longitude, latitude}}{longitude and latitude of the known geographical location}
}
\description{
Given a known geographical location, this function downloads the corresponding 
raw text file of observations for the closest station from 
"ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/gsn/", and converts the relevant 
average, maximum, minimum daily temperature into a dataframe.
This function also reports the distance between the given geographical location and 
the corresponding meteostation.
}
\examples{
require(TuktuTools)
a <- getWeatherAtLocation(-152.5, 68.25, 
                          start = 2010, end = 2022)

# look at stations
data("gsn_stations")
gsn_stations \%>\% head
gsn_stations \%>\% subset(station \%in\% a$Station)

# plot avg temperature by year
require(ggplot2)

a \%>\% subset(Month == 7,) \%>\% arrange(Day) \%>\% 
  ggplot(aes(Day, Tavg, col = Station)) + 
  geom_ribbon(aes(ymin = Tmin, ymax = Tmax)) + 
  geom_point() + geom_path() + facet_wrap(.~Year)

# map stations in Alaska region where lon < -130 & lat > 60
require(sf)
gsn.sf <- st_as_sf(gsn_stations \%>\% subset(lon < -130 & lat > 60), 
                   coords = c("lon", "lat")) \%>\% st_set_crs(4326)

mapview::mapview(gsn.sf[,"station"])
}
