% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnllM.R
\name{mnll2M}
\alias{mnll2M}
\alias{mnll3M}
\title{Function that minimises the likelihood of the no calf, calf and calf death models for a given individual}
\usage{
mnll2M(df, int, kcons)

mnll3M(df, int, kcons)
}
\arguments{
\item{df}{dataframe from get.speed function}

\item{int}{integer value indicating the minimum number of days between
the beginning of the time series and the first BP (calf birth),
the last BP (calf death) and the end of the time series,
and between the two BPs.
The main reason for this constraint is that a minimum number of data points are required in
each section of the time-series to be able to estimate model parameters.
We recomanded 9 relocations, thereby 3 days.}

\item{kcons}{vector of the minimum and maximum time it takes the female to recover normal speed (in days)
to recover her normal movement (in days).}
}
\value{
# returns a list of three objects,
which contains the MLE values of all parameters and the mnll, AIC coefficient
of all no calf, calf and calf death models models
}
\description{
\itemize{
 \item mnll2M minimize the negative log-likelihood of the no calf and calf models only
 \item mnll3M minimize the negative log-likelihood of all the three models
}
}
\section{Functions}{
\itemize{
\item \code{mnll3M}: minimize the negative log-likelihood of all the three models
}}

\references{
DeMars, C., M. Auger-Méthé, U. Schlägel, S. Boutin, (Published online) Inferring Parturition and Neonate Survival from Movement Patterns of Female Ungulates. Ecology and Evolution. DOI: 10.1002/ece3.785
}
