% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllcalfdeath.R
\name{nll.calfdeath}
\alias{nll.calfdeath}
\title{Likelihood of Female which gave birth but the calf died}
\usage{
nll.calfdeath(df, BP1, BP2, kcons, PlotMe = FALSE)
}
\arguments{
\item{df}{dataframe from get.speed function}

\item{BP1}{Break point corresponding to calf birth (in days from the begining of the dataset)}

\item{BP2}{Break point corresponding to calf death (in days from the begining of the dataset)}

\item{kcons}{vector of the minimum and maximum time it takes the female to recover normal speed (in days)}

\item{PlotMe}{if PlotMe = TRUE, a plot of the speed in function to the date and time will be drawn
for each individual with the line of the speed estimated by the function with depending of BP1, BP2 and k}
}
\value{
Returns the negative log-likelihood of this model and the recovery time (in days)
}
\description{
nll.calfdeath calculates the negative log-likelihood of a female with a calf that died
}
\details{
This model assumes that the mean speed is constant prior to the birth of a calf and after calf's death

Once the calf is born the speed decrease close to 0.
Then the speed increases slowly until the calf dies,
then increases imediatly to reach the initial speed before parturition after the calf is dead

Once the mean speed reaches the value from before the birth of the calf,
it remains at this value.
}
\references{
DeMars, C., M. Auger-Méthé, U. Schlägel, S. Boutin, (Published online) Inferring Parturition and Neonate Survival from Movement Patterns of Female Ungulates. Ecology and Evolution. DOI: 10.1002/ece3.785
}
