% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllpost.R
\name{nll.post}
\alias{nll.post}
\title{Likelihood of Female with a calf}
\usage{
nll.post(
  par = c(log.beta.calf = NA, recovery = NA),
  alpha.mean,
  beta.mean,
  alpha.calf,
  dhours.b,
  speed.b
)
}
\arguments{
\item{alpha.mean}{shape of the Gamma distribution function used to calculate speed before birth}

\item{beta.mean}{rate of the Gamma distribution function used to calculate speed before birth}

\item{alpha.calf}{parameter used to calculate the shape of the Gamma distribution function used to calculate speed during recovery}

\item{dhours.b}{vector of the difference in hour between each speed value and the first one}

\item{speed.b}{vector of the speed after birth}

\item{beta.calf}{parameter used to calculate the rate of the Gamma distribution function used to calculate speed during recovery}

\item{recovery}{the recovery time is a possibility}
}
\value{
Returns the negative log-likelihood of this model given the data SL
}
\description{
nll.post calculates the negative log-likelihood of the section of the time series
when the female is with a calf in both cases where the calf survived and the one where the calf died.
It is used in part to estimate k.
k represents the time it takes, in days, for the female to recover her normal movement.
}
\details{
nll.post is minimised within both models where the calf survived and the calf died (nll.calf and nll.calfdeath).

The function assumes an increase of the speed of the female's movement after the birth of a calf.
}
\references{
DeMars, C., M. Auger-Méthé, U. Schlägel, S. Boutin, (Published online) Inferring Parturition and Neonate Survival from Movement Patterns of Female Ungulates. Ecology and Evolution. DOI: 10.1002/ece3.785
}
