% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parturitionmodel.R
\name{parturition.model}
\alias{parturition.model}
\title{Parturition timing function}
\usage{
parturition.model(
  df,
  int,
  kcons,
  PlotIt = FALSE,
  saveplot = FALSE,
 
    CRS = "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
)
}
\arguments{
\item{df}{a dataframe containing the speed between subsequent relocations of individuals,
the coordinates of the relocations (in metric system)
the date and time (Time) of each speed value and the difference in hours 
between each speed value and the first one
and a vector of the animal identifiant (ID).
See ?get.speed for more information on the Data requirements}

\item{int}{integer value indicating the minimum number of days between
the beginning of the time series and the first BP (calf birth),
the last BP (calf death) and the end of the time series,
and between the two BPs.
The main reason for this constraint is that a minimum number of data points are required in
each section of the time-series to be able to estimate model parameters.
We recomanded 9 relocations, thereby 3 days.}

\item{kcons}{vector of the minimum and maximum time it takes the female to recover normal speed (in days)}

\item{PlotIt}{if PlotIt = TRUE the function will draw a plot of the speed in function to the date
with the prediction line ot the best model selected (by 'AIC'),
among no-calf, calf, calf death based on the actual speed of the female in function of the date}

\item{saveplot}{if saveplot = TRUE, the plot of the best model will be saved}

\item{CRS}{the coordinates projection (default is Canada Lambert Conformal Conic:
"+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs")}
}
\value{
returns list of
the AIC of the 3 models (i.e. no calf model, calf model and calf death model),
the best model based on AIC, the dates of calf and death if any Recovery time (in days) for each calf model,
Par, the estimated parameters of the best model based on AIC and
a summary of the results with the best model based on AIC, the calving date if any,
the mortality date of the calf (if any), the recovery time (in days, if any),
and the x and y calving locations
}
\description{
Function that determine calving status of a female (i.e. no calf, with a calf, calf lost)
the calving date and the calf death date (if any).
This is an adaptation of the individual based method developped by Demars et al. (2013),
which has proven good reliability to estimate calving status and calving date for females
from the Western Arctic Herd in a previous study (Cameron et al. 2018).
However, we adapted this method to be able to infer parturition based on the female speed through time.
}
\details{
All models assume speed follows a Gamma distribution and differ in two parameters: shape
and scale, which correspond to \eqn{mean(speed)^2 / var(speed)}
and \eqn{var(speed) / mean(speed)}, respectively.
The mean speed is thus equal to \eqn{shape * scale}.
\itemize{
\item For the model representing females that do not calve: the mean speed remains constant
through the entire calving period.
\item For the model representing females that had a calf who survived 4 weeks after birth:
the mean speed is constant before calving, then abruptly drops for calving, creating a break point.
After calving, the mean speed increases progressively following:
\eqn{ (shape.calving * (mean(shape) - shape.calving) / k * time) * (scale.calving * (mean(scale) - scale.calving) / k) * time)}
where k is the time, defined in days, required for the calf to achieve adult movement rates.
\item For the model representing females losing calves: there is an abrupt change in the slope of
the post-calving increase, creating a second break point after which the mean speed immediately
recovers its pre-calving value.}
The models therefore differ in their number of parameters to estimate:
the no calf model has two – shape and scale;
the calf model has four – shape, scale, k and calving date;
and the calf death model has five – shape, scale, k, calving and calf death.
We discriminated among models using Akaike’s Information Criterion (AIC)
with the best model being the one with the lowest AIC value.

The visualization portion of this code is adapted directly from Matt Cameron and supplementary materials in Cameron et al. 2018.
}
\references{
DeMars, C., M. Auger-Méthé, U. Schlägel, S. Boutin, (Published online) Inferring Parturition and Neonate Survival from Movement Patterns of Female Ungulates. Ecology and Evolution. DOI: 10.1002/ece3.785
Cameron, M.D., Joly, K., Breed, G.A., et al. (2018). Movement-based methods to infer parturition events in migratory ungulates. Canadian Journal of Zoology. 96:1187–1195. https://doi.org/10.1139/cjz-2017-0314
}
