% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parturitionmodel2.R
\name{parturition.model2}
\alias{parturition.model2}
\title{Parturition timing function considering 2 models}
\usage{
parturition.model2(
  df,
  int,
  kcons,
  method,
  PlotIt = FALSE,
  saveplot = FALSE,
  CRS =
    "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
)
}
\arguments{
\item{df}{a dataframe containing the speed between subsequent relocations of individuals,
the coordinates of the relocations (in metric system)
the date and time oh each speed value and the difference in hours 
between each speed value and the first one
and a vector of the animal-year identifiant.
See ?get.speed for more information on the Data requirements}

\item{int}{integer value indicating the minimum number of days between
the beginning of the time series and the first BP (calf birth),
and BP and the end of the time series.
The main reason for this constraint is that a minimum number of data points are required in
each section of the time-series to be able to estimate model parameters.
We recomanded 9 relocations, thereby 3 days.}

\item{kcons}{vector of the minimum and maximum time it takes the female to recover normal speed (in days)}

\item{PlotIt}{if PlotIt = TRUE the function will draw a plot of the speed in function to the date
with the prediction line ot the best model selected (by 'AIC'),
among no-calf, calf, calf death based on the actual speed of the female in function of the date}

\item{saveplot}{if saveplot = TRUE, the plot of the best model will be saved}

\item{CRS}{the coordinates projection (default is Canada Lambert Conformal Conic:
"+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs")}
}
\value{
returns list of
the AIC of the 2 models (i.e. no calf model and calf model),
the best model based on AIC, the dates of calf and death if any Recovery time (in days) for each calf model,
Par, the estimated parameters of the best model based on AIC and
a summary of the results with the best model based on AIC, the calving date if any,
and the recovery time (in days, if any),
a Z-score for the calving date, as a deviation from average calving date from Cameron et al. 2018
and the x and y calving locations
}
\description{
Function that determine calving status of a female (i.e. no calf, with a calf)
the calving date and the recovery time (if any).
This is an adaptation of the individual based method developped by Demars et al. (2013),
which has proven good reliability to estimate calving status and calving date for females
from the Western Arctic Herd in a previous study (Cameron et al. 2018).
However, we adapted this method to be able to infer parturition based on the female speed through time.
}
\details{

}
\references{
DeMars, C., M. Auger-Méthé, U. Schlägel, S. Boutin, (Published online) Inferring Parturition and Neonate Survival from Movement Patterns of Female Ungulates. Ecology and Evolution. DOI: 10.1002/ece3.785
Cameron, M.D., Joly, K., Breed, G.A., et al. (2018). Movement-based methods to infer parturition events in migratory ungulates. Canadian Journal of Zoology. 96:1187–1195. https://doi.org/10.1139/cjz-2017-0314
}
