% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parturitionplot.R
\name{parturition.plot}
\alias{parturition.plot}
\title{Parturition model plot function}
\usage{
parturition.plot(df, int, kcons, bestmodel, method)
}
\arguments{
\item{df}{a dataframe containing the speed between subsequent relocations of an individual,
the coordinates of the relocations (in metric system)
the date and time oh each speed value and t
he difference in hours between each speed value and the first one
and a vector of the animal-year identifiant.
See ?get.speed for more information on the Data requirements}

\item{int}{integer value indicating the minimum number of days between
the beginning of the time series and the first BP (calf birth),
the last BP (calf death) and the end of the time series,
and between the two BPs.
The main reason for this constraint is that a minimum number of data points are required in
each section of the time-series to be able to estimate model parameters.
We recomanded 9 relocations, thereby 3 days.}

\item{kcons}{vector of the minimum and maximum time it takes the female to recover normal speed (in days)}

\item{bestmodel}{the selected model by the parturition.model function, for this individual among: 'nocalf',
'calf', 'calfdeath'}
}
\value{
The function returns a plot of the speed pattern and the fit of the model specified by bestmodel
as well as the plot of the track of the individual with the calving location (if bestmodel is equal to calf or calfdeath)
}
\description{
Function to plot speed pattern of a female with the fitted values estimated by
the parturition.model given the chosen bestmodel.
among 'nocalf', 'calf', 'calfdeath'
See ?parturition.model for more information about the method
}
\details{
The visualization portion of this code is adapted directly from the supplementary materials in Cameron et al. 2018.
}
\references{
DeMars, C., M. Auger-Méthé, U. Schlägel, S. Boutin, (Published online) Inferring Parturition and Neonate Survival from Movement Patterns of Female Ungulates. Ecology and Evolution. DOI: 10.1002/ece3.785
Cameron, M.D., Joly, K., Breed, G.A., et al. (2018). Movement-based methods to infer parturition events in migratory ungulates. Canadian Journal of Zoology. 96:1187–1195. https://doi.org/10.1139/cjz-2017-0314
}
