% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCalving.R
\name{plotCalving}
\alias{plotCalving}
\title{Plotting calving}
\usage{
plotCalving(df, int, kcons, bestmodel)
}
\arguments{
\item{df}{a dataframe containing the speed between subsequent relocation of ONE individual,
obtained using `getSpeed`. 
See ?get.speed for more information on the Data requirements}

\item{int}{integer value indicating the minimum number of days between
the beginning of the time series and the first BP (calf birth),
the last BP (calf death) and the end of the time series,
and between the two BPs.
The main reason for this constraint is that a minimum number of data points are required in
each section of the time-series to be able to estimate model parameters.
We recomanded 9 relocations, thereby 3 days.}

\item{kcons}{vector of the minimum and maximum time it takes the female to recover normal speed (in days)}

\item{bestmodel}{the selected model by the parturition.model function, for this individual among: 'nocalf',
'calf', 'calfdeath'}
}
\value{
The function returns a plot of the speed pattern and the fit of the model specified by bestmodel
as well as the plot of the track of the individual with the calving location (if bestmodel is equal to calf or calfdeath)
}
\description{
Function to plot speed pattern of a female with the fitted values estimated by
the parturition.model given the chosen bestmodel.
among 'nocalf', 'calf', 'calfdeath'
See ?estimateCalving for more information about the method
}
\details{
The visualization portion of this code is adapted directly from the 
supplementary materials in Cameron et al. 2018.
}
\examples{
require(TuktuTools)

data(caribou)

# remove potential outliers and keep only females
caribou.cleaned <- removeOutliers(caribou) \%>\% subset(outlier == "FALSE" & sex == "f")

# keep data during calving period (i.e., May 19 to July 7), no more than 3 consecutive days of missing data
# and at least 1 fix per day
caribou.prepped <- prepData(caribou.cleaned, start = "05-19", end = "07-07", 
                            nfixes = 1, dayloss = 3)

# get movement rate for the females
caribou.mr <- getSpeed(caribou.prepped)
head(caribou.mr)

# estimate calving with the no calf and calf model
parturitions <- estimateCalving(caribou.mr, int=3, kcons =c(5,21), models = "full", drawplot = TRUE)
head(parturitions)

# get plot for one individual
table(caribou.mr$ID, caribou.mr$Year) # take Vixen in 2008

# what is the identified model for Vixen in 2008?
parturitions$results[parturitions$results$ID == "Vixen" & 
                       parturitions$results$Year == 2008,] # "calfdeath"

plotCalvingRange(df = caribou.mr[caribou.mr$ID == "Vixen" & 
                                   caribou.mr$Year == 2008,], int = 3,
               kcons = c(5,21), bestmodel = "calfdeath")
}
\references{
DeMars, C., M. Auger-Méthé, U. Schlägel, S. Boutin, (Published online) Inferring Parturition and Neonate Survival from Movement Patterns of Female Ungulates. Ecology and Evolution. DOI: 10.1002/ece3.785
Cameron, M.D., Joly, K., Breed, G.A., et al. (2018). Movement-based methods to infer parturition events in migratory ungulates. Canadian Journal of Zoology. 96:1187–1195. https://doi.org/10.1139/cjz-2017-0314
}
