% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Process a timeseries for the period of interest}
\usage{
prepData(df, start, end, nfixes = 1e-04, dayloss = Inf, restrictive = FALSE)
}
\arguments{
\item{df}{a data frame containing columns: ID as individual identifiant,
x and y: relocations of individuals (in N Canada Lambert Conformal Conic)
Time: date and time vector of the relocation (of class POSIXct)}

\item{start}{starting date of the period of interest as a character in the form "mm-dd"
(example for the 19th of May: "05-19")}

\item{end}{end date of the period of interest as a character in the form "mm-dd" (excluded)
(example for the 8th of July: "07-08")}

\item{nfixes}{individuals with fewer than nfixes observations per day over the period of of interest are removed}

\item{dayloss}{maximum number of consecutive days with missing locations
(for example, if an individual has loss signal for more than 3 consecutive days, 
it will be excluded from the dataset)}

\item{restrictive}{if TRUE, only individuals having monitoring for the entire period 
are kept}
}
\value{
a processed dataframe
}
\description{
Processes and filters caribou movement data by 
(a) trimming to analysis period of interest; for example, in the parturition analysis
dates were limited to May 19 to July 7 ; 
(b) guaranteeing a minimum number of location fixes a day (e.g. 1 for the parturiton analysis), and removing 
individuals that have too few; and 
(c) removing individuals with a data gap greater than some minimum threshold (e.g. 3 days for the parturiton analysis).
}
\examples{
require(TuktuTools)

data(caribou)

# keep data during calving period (i.e., May 19 to July 7), no more than 3 consecutive days of missing data
# and at least 1 fix per day
prepped.caribou <- prepData(caribou.cleaned, start = "05-19", end = "07-07", nfixes = 1, dayloss = 3)
}
