% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_moveNWT.R
\name{process_moveNWT}
\alias{process_moveNWT}
\title{Process Movebank data from NWT}
\usage{
process_moveNWT(
  movedata,
  CRS =
    "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs",
  mylogin
)
}
\arguments{
\item{movedata}{a Move object, typically obtained with the \code{\link{getMovebankData}} command.}

\item{CRS}{coordinate reference system, in any format suitable for \code{\link{st_transform}}}

\item{moveLogin}{a movebankLogin object as: movebankLogin(username="MyUsername", password="MyPassword")}
}
\value{
a dataframe with columns:
\describe{
  \item{ID}{ID of animal}
  \item{nickname}{"nick name" of animal from movebank data}
  \item{DateTime}{Date and time in POSIXct}
  \item{Year}{year}
  \item{Lon,Lat}{Longitude and Latitude (WGS84)}
  \item{x,y}{coordinates in the given projection (default: Canada Lambert Conformal Conic)}
  \item{sex}{Sex of the individual (f: female, m: male)}
  \item{study.site}{Individual's herd assignment}
}
}
\description{
This function takes a move object from Movebank from the GNWT data sets and returns it in a format
suitably for many of the analyses in this package.
}
\examples{
# to generate the bluenose data
\dontrun{
bluenose.move <- getMovebankData("GNWT Sahtu Barren Ground Caribou: Bluenose-East",
login = mylogin, removeDuplicatedTimestamps=TRUE)
bluenose <- process_moveNWT(bluenose.move)}
}
