% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_tracks.R
\name{scan_tracks}
\alias{scan_tracks}
\title{Scan Multiple Tracks at a Glance}
\usage{
scan_tracks(
  sf.list,
  cols = 1:length(sf.list),
  distance.df = NULL,
  plotdistance = FALSE,
  d.col = "darkgrey",
  legend = FALSE,
  threshold = 1000,
  ...
)
}
\arguments{
\item{sf.list}{list of simple feature tracks}

\item{cols}{vector of colors}

\item{distance.df}{distance data frame (output of \code{\link{getPairwiseDistances}})}

\item{plotdistance}{whether to plot pairwise distances (only works for list of length 2)}

\item{d.col}{col of distances}

\item{legend}{whether or not to add a legend}
}
\value{
If plotdistance = TRUE, returns the pairwise distance data frame.
}
\description{
Takes list of individual (simple feature) tracks, and plots their tracks.
Useful, e.g., for comparing proximity of animals. Includes a feature which computes
and plots the distance between daily mean locations.
}
\examples{
data(bathurst)
b.subset <- bathurst \%>\% subset(Year == 2014)
b.list <- dlply(b.subset, "ID", st_as_sf)


# basic quicky implementation
require(gplots)
palette(rich.colors(length(b.list)))
scan_tracks(b.list)

# with pairwise distances of a subset of indviduals
b.distances <- scan_tracks(b.list[1:6], cols = rich.colors(6), legend = TRUE,
            plotdistance = TRUE, threshold = 1e3)

head(b.distances)
}
