% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBioYear.R
\name{addBioYear}
\alias{addBioYear}
\title{Add Biological Year to data}
\usage{
addBioYear(df, bioStart)
}
\arguments{
\item{df}{a dataframe with columns:
`Time` Date and time in POSIXct}

\item{bioStart}{biological year start date, in day of year format \code{\link{yday}}}
}
\value{

}
\description{
This function adds a column designating biological year starting on a user-designated date 
e.g. 1 June (day of year 152) for caribou
}
\examples{
data("caribou")

head(caribou)

# bioStart is yday 152 (01 June)

caribou <- addBioYear(caribou, bioStart = 152)
head(caribou)
}
