% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{caribou}
\alias{caribou}
\title{Movement data of barren-ground caribou}
\format{
data frame with the following columns: 
#' \describe{
  \item{id}{ID of animal}
  \item{start}{Date of beginning of collaring}
  \item{end}{Date of death or censoring}
  \item{fate}{One of "dead", or "censored"}
}

name of 4 individuals (ID), the sex, the coordinates 
of the GPS locations (Lon,Lat in WGS84 and x,y in Canada Lambert Conformal Conic) 
and date and time (Time) of the GPS locations. These are real data courtesy of NWT
but anonymized and shifted in time, longitude and latitude.
}
\source{
GNWT Department of Environment and Natural Resources
}
\usage{
data(caribou)
}
\description{
Movement data of barren-ground caribou
}
\examples{
data(caribou)
head(caribou)

# plotting locations on a base graphics map
require(mapdata)
maps::map("worldHires", xlim = c(-135,-120), ylim=c(62,67), col = "grey")
ids <- levels(caribou$ID)
for(i in 1:length(ids))
  with(caribou \%>\% subset(ID == ids[i]), lines(Lon, Lat, col = i))

# plotting points with mapview
caribou.sf <- st_as_sf(caribou \%>\% mutate(Lat = Lat, Lon = Lon), 
                       coords = c("Lon","Lat")) \%>\% st_set_crs(4326)
mapview::mapview(caribou.sf[,"ID"])

# converting to lines
}
\keyword{data}
