% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{caribou}
\alias{caribou}
\title{Movement data of barren-ground caribou}
\format{
data frame with the name of the individuals (ID), the sex, the coordinates of the GPS locations (Lon,Lat in WGS84 and x,y in Canada Lambert Conformal Conic) and date and time (Time) of the GPD locations
}
\source{
GNWT
}
\usage{
data(caribou)
}
\description{
Movement data of barren-ground caribou
}
\examples{
data(caribou)
head(caribou)

}
\keyword{data}
