% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCrossingDay.R
\name{getCrossingTime}
\alias{getCrossingTime}
\title{Get migration arrival from calving ground}
\usage{
getCrossingTime(df, movedata, percent, method)
}
\arguments{
\item{df}{a dataframe containing the ID_Year, the herd and lon and lat of the calving event (in WGS 84)}

\item{movedata}{a simple feature containing the movement data for all individuals in df, in WGS84 (crs=4326) (with column ID_Year and DateTime of the relocation)}

\item{percent}{the percent to be used to estimate the calving ground}

\item{method}{the method to be used to estimate the calving ground of the herd (prefrerred 'LSCV')}
}
\value{
a dataframe with the ID_Year and the day the individual crossed the boundaries
of the calving ground (estimated based on the method and percent defined in the parameters)
}
\description{
Function to determine migration arrival from the calving ground
(estimated based on the individuals'calving locations).
}
\details{
The function estimates the calving ground of each herd, based on calving locations sepecified given in df
Then it determines the migration arrival time as the time at which a given individual crossed the boundaries of the herd
calving ground (i.e. the first relocation inside the calving range)
}
