% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDailyMean.R
\name{getDailyMean}
\alias{getDailyMean}
\title{Calculate Daily Mean}
\usage{
getDailyMean(mysf)
}
\arguments{
\item{mysf}{a simple feature subseted to a single ID.}
}
\value{
A simple feature with the same CRS containing daily mean locations.
}
\description{
Calculate Daily Mean
}
