% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDailyMean.R
\name{getDailyMean}
\alias{getDailyMean}
\title{Calculate Daily Mean}
\usage{
getDailyMean(df)
}
\arguments{
\item{df}{a data frame with the ID, the Time (as date and time), the x and y coordinates in metric system, 
and the Lon and Lat coordinates in WGS84}
}
\value{
A data frame with the daily mean locations
}
\description{
Important note - requires a column called "DateTime"
}
\examples{
require(TuktuTools)

data(caribou)

daily_mean <- getDailyMean(caribou)
head(daily_mean)
}
