% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHerdAssignment.R
\name{getHerdAssignment}
\alias{getHerdAssignment}
\title{Get herd assignment using spatial clustering}
\usage{
getHerdAssignment(df, K, PlotIt = TRUE)
}
\arguments{
\item{df}{a dataframe containing x and y coordinates (in a metric system) with Time and the ID of individuals}

\item{K}{number of herds (can be determined using getKnumber function in this package)}

\item{plotIt}{if TRUE, plot the GPS location with herd assignment}
}
\value{
the dataframe with GPS location and a cluster column with
a number comprised between 1 and K, corresponding to the herd assignment.
}
\description{
Function to identify different herds/populations using K-means analysis on GPS locations of individuals,
using the average location for each individual and year.
}
\examples{
require(TuktuTools)

data(barrenground)

# Try the getHerdAssignment on the pool of individuals we have, with 2 clusters
assignment <- getHerdAssignment(barrenground, K = 2)

# we see that one individuals being further from the other ones
# just out of curiosity, let's try with K = 3
assignment <- getHerdAssignment(barrenground, K = 3)

# the returned dataframe now contains one column with the cluster
head(assignment)
}
