% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKernelUD.R
\name{getKernelUD}
\alias{getKernelUD}
\title{get Kernel UD}
\usage{
getKernelUD(my.sf, percent = 95, grid = 200, ...)
}
\arguments{
\item{my.sf}{a simple feature point collection}

\item{percent}{percent kernel}

\item{grid}{grid points (default 200, higher than adehabitat's)}

\item{...}{additional parameters to pass to \link{kernelUD}}
}
\description{
wrapper for \link{kernelUD} in adehabitatLT that returns a given UD
as a simple feature polygon from a simple feature point collection
}
\examples{
data("barrenground")

b <- subset(barrenground, ID == ID[1]) \%>\% arrange(DateTime)
with(b, plot(x,y, type = "o"))

# convert to simple feature

b.sf <- st_as_sf(b, coords = c("Lon","Lat")) \%>\% st_set_crs(4326)
b.ud.95 <- getKernelUD(b.sf)
b.ud.50 <- getKernelUD(b.sf, 50)
m1 <- mapview(b.sf)
m2 <- mapview(b.ud.95, map = m1@map, col.regions = "red")
mapview(b.ud.50, map = m2@map, col.regions = "red")
}
