% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKernelUD.R
\name{getKernelUD}
\alias{getKernelUD}
\title{get Kernel UD}
\usage{
getKernelUD(my.sf, percent = 95, grid = 200, ...)
}
\arguments{
\item{my.sf}{a simple feature point collection}

\item{percent}{percent kernel}

\item{grid}{grid points (default 200, higher than adehabitat's)}

\item{...}{additional parameters to pass to \link{kernelUD}}
}
\description{
wrapper for \link{kernelUD} in adehabitatLT that returns a given UD
as a simple feature polygon from a simple feature point collection
}
