% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnll2M.R
\name{mnll2M}
\alias{mnll2M}
\title{Function that minimises the likelihood of the nll.post and nll.calf models for a given individual}
\usage{
mnll2M(df, int, kcons)
}
\arguments{
\item{df}{dataframe from get.speed function}

\item{int}{integer value indicating the minimum number of days between
the beginning of the time series and BP (calf birth),
and BP and the end of the time series.
The main reason for this constraint is that a minimum number of data points are required in
each section of the time-series to be able to estimate model parameters.
We recomanded 9 relocations, thereby 3 days.}

\item{kcons}{vector of the minimum and maximum time it takes the female to recover normal speed (in days)
to recover her normal movement (in days).}
}
\value{
# returns a list of three objects,
which contains the MLE values of all parameters and the mnll, AIC coefficient
of the no calf or calf model
}
\description{
Function that minimises the likelihood of the nll.post and nll.calf models for a given individual
}
\references{
DeMars, C., M. Auger-Méthé, U. Schlägel, S. Boutin, (Published online) Inferring Parturition and Neonate Survival from Movement Patterns of Female Ungulates. Ecology and Evolution. DOI: 10.1002/ece3.785
}
