% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllcalf.R
\name{nll.calf}
\alias{nll.calf}
\title{Likelihood of Female giving birth}
\usage{
nll.calf(df, BP1, kcons, PlotMe = FALSE)
}
\arguments{
\item{df}{dataframe from get.speed function}

\item{kcons}{vector of the minimum and maximum time it takes the female to recover normal speed (in days)}

\item{PlotMe}{if PlotMe = TRUE, a plot of the speed in function to the date and time will be drawn
for each individual with the line of the speed estimated by the function with depending of BP}

\item{BP}{Break point corresponding to calf birth (in days from the begining of the dataset)}
}
\value{
Returns the negative log-likelihood of this model and the recovery time (in days)
}
\description{
nll.calf calculates the negative log-likelihood of a female with a calf that survived.
}
\details{
This model assumes that the mean speed is constant prior to the birth of a calf.

Once the calf is born the speed decrease close to 0.
Then the speed increases to reach the same speed as prior to calf birth.
The increase depends on the time it takes to recover the normal movement (in days),
which is represented by k.

Once the mean speed reaches the value from before the birth of the calf,
it remains at this value.
}
\references{
DeMars, C., M. Auger-Méthé, U. Schlägel, S. Boutin, (Published online) Inferring Parturition and Neonate Survival from Movement Patterns of Female Ungulates. Ecology and Evolution. DOI: 10.1002/ece3.785
}
