% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeoutliers.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove almost duplicated timestamp}
\usage{
removeOutliers(
  df,
  steps = 10,
 
    CRS = "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
)
}
\arguments{
\item{df}{a data frame containing columns: ID as individual identifiant,
x and y: relocations of individuals (metric system specified in CRS)
Time: vector (of class POSIXct)}

\item{steps}{if specified, the number of cleaning steps to be performed (default is 10)}

\item{CRS}{the coordinates projection (default is Canada Lambert Conformal Conic:
"+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs")}
}
\value{
a list with a dataframe without potential 'outliers' and a dataframe of outliers
}
\description{
Function to remove almost duplicated timestamp.
It appears that sometimes, the GPS device relocate at 1 or 2
minute interval. However, the GPS devices are configured to
relocate every day, 8-hours or every hour for some
individuals, at fix hours (e.g. 0:00 am, 8:00 am or 4:00 pm).
In addition, it is biologically impossible that the animal
moves several kilometers in one minute. Thus, the relocations
having a time interval in the order of minutes are more prone
to be 'outliers'.
}
\details{
I noted that sometimes there could be several subsequent "outliers"
thus the cleaning function is repeated several times until there are no "outliers" anymore
}
