% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeoutliers.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove (near) duplicated timestamp}
\usage{
removeOutliers(df, steps = 10, max.speed = 50, min.interval = 2)
}
\arguments{
\item{df}{a data frame containing columns: ID as individual identifiant,
x and y: relocations of individuals (in a metric system)
Time: vector (of class POSIXct)}

\item{steps}{if specified, the number of cleaning steps to be performed (default is 10)}

\item{max.speed}{threshold (in km/hour) for removing outliers.  Default: 50 km/hr}

\item{min.interval}{threshold (in minutes) for mininum intervals between location. Default: 2 min.}
}
\value{
The original data frame with an additional column labeled "outlier".
}
\description{
Function to remove almost duplicated timestamps. Occasionally the GPS device 
transmits locations at 1 or 2 minute intervals despite the duty cycle being 
set to 1 day, 8 hours, or - at most - 2 hours.  This leads notably to highly 
unrealistic speeds (e.g. several kilometers in one minute). This function 
flags locations with very short intervals or unrealistic speeds as outliers.
}
\examples{
require(TuktuTools)

data(caribou)

# flag outliers
caribou.cleaned <- removeOutliers(caribou)

head(caribou.cleaned)

table(caribou.cleaned$outlier)

# visualize 
ggplot(data=caribou.cleaned, aes(x=x, y = y)) + geom_path() + 
  geom_point(data= caribou.cleaned \%>\% subset(outlier == "TRUE"), aes(colour=factor(outlier))) + 
  facet_wrap(~ID, scales = "free")


which(caribou.cleaned$outlier == "TRUE") # [1] 12263 12326 13394 13689

# check those locations
caribou.cleaned[12262:12264,] # one minute interval

caribou.cleaned[12325:12327,] # one minute interval

caribou.cleaned[13393:13395,] # one minute interval

caribou.cleaned[13687:13690,] # one minute interval
}
