% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_tracks.R
\name{scan_tracks}
\alias{scan_tracks}
\title{Scan Multiple Tracks at a Glance}
\usage{
scan_tracks(
  x,
  id.col = "ID",
  time.col = "Time",
  colors = NULL,
  distance.df = NULL,
  plotdistance = FALSE,
  d.col = "darkgrey",
  legend = FALSE,
  threshold = 1000,
  legend.pos = "bottomleft",
  ...
)
}
\arguments{
\item{id.col}{name of ID column}

\item{time.col}{name of Time column}

\item{distance.df}{distance data frame (output of \code{\link{getPairwiseDistances}})}

\item{plotdistance}{whether to plot pairwise distances (only works for list of length 2)}

\item{d.col}{col of distances}

\item{legend}{whether or not to add a legend}

\item{...}{additional parameters to pass to legend function}

\item{sf.list}{list of simple feature tracks}

\item{cols}{vector of colors}
}
\value{
If plotdistance = TRUE, returns the pairwise distance data frame.
}
\description{
Takes list of individual (simple feature) tracks, and plots their tracks.
Useful, e.g., for comparing proximity of animals. Includes a feature which computes
and plots the distance between daily mean locations.
}
